+++ app/javascript/mastodon/features/account_timeline/containers/header_container.js
@@ -33,6 +33,7 @@ const makeMapStateToProps = () => {
 
   const mapStateToProps = (state, { accountId }) => ({
     account: getAccount(state, accountId),
+    domain: state.getIn(['meta', 'domain']),
   });
 
   return mapStateToProps;
