+++ app/javascript/styles/mastodon/containers.scss
@@ -677,6 +677,7 @@
           color: $darker-text-color;
           text-decoration: none;
           padding: 15px;
+          font-weight: 500;
 
           strong {
             font-weight: 700;
